+++ CodeEdit/Features/LSP/Registry/PackageManagerProtocol.swift
@@ -23,7 +23,7 @@ protocol PackageManagerProtocol {
 
 extension PackageManagerProtocol {
     /// Creates the directory for the language server to be installed in
-    internal func createDirectoryStructure(for packagePath: URL) throws {
+    func createDirectoryStructure(for packagePath: URL) throws {
         let decodedPath = packagePath.path.removingPercentEncoding ?? packagePath.path
         if !FileManager.default.fileExists(atPath: decodedPath) {
             try FileManager.default.createDirectory(
@@ -35,56 +35,18 @@ extension PackageManagerProtocol {
     }
 
     /// Executes commands in the specified directory
-    internal func executeInDirectory(in packagePath: String, _ args: [String]) async throws -> [String] {
+    func executeInDirectory(in packagePath: String, _ args: [String]) async throws -> [String] {
         return try await runCommand("cd \"\(packagePath)\" && \(args.joined(separator: " "))")
     }
 
     /// Runs a shell command and returns output
-    internal func runCommand(_ command: String) async throws -> [String] {
+    func runCommand(_ command: String) async throws -> [String] {
         var output: [String] = []
         for try await line in shellClient.runAsync(command) {
             output.append(line)
         }
         return output
     }
-}
-
-enum PackageManagerError: Error {
-    case packageManagerNotInstalled
-    case initializationFailed(String)
-    case installationFailed(String)
-    case invalidConfiguration
-}
-
-enum RegistryManagerError: Error {
-    case invalidResponse(statusCode: Int)
-    case downloadFailed(url: URL, error: Error)
-    case maxRetriesExceeded(url: URL, lastError: Error)
-    case writeFailed(error: Error)
-}
-
-/// Package manager types supported by the system
-enum PackageManagerType: String, Codable {
-    /// JavaScript
-    case npm
-    /// Rust
-    case cargo
-    /// Go
-    case golang
-    /// Python
-    case pip
-    /// Ruby
-    case gem
-    /// C#
-    case nuget
-    /// OCaml
-    case opam
-    /// PHP
-    case composer
-    /// Building from source
-    case sourceBuild
-    /// Binary download
-    case github
 }
 
 /// Generic package source information that applies to all installation methods.
@@ -130,50 +92,5 @@ struct PackageSource: Equatable, Codable {
     enum GitReference: Equatable, Codable {
         case tag(String)
         case revision(String)
-    }
-}
-
-/// Installation method enum with all supported types
-enum InstallationMethod: Equatable {
-    /// For standard package manager installations
-    case standardPackage(source: PackageSource)
-    /// For packages that need to be built from source with custom build steps
-    case sourceBuild(source: PackageSource, command: String)
-    /// For direct binary downloads
-    case binaryDownload(source: PackageSource, url: URL)
-    /// For installations that aren't recognized
-    case unknown
-
-    var packageName: String? {
-        switch self {
-        case .standardPackage(let source),
-             .sourceBuild(let source, _),
-             .binaryDownload(let source, _):
-            return source.pkgName
-        case .unknown:
-            return nil
-        }
-    }
-
-    var version: String? {
-        switch self {
-        case .standardPackage(let source),
-             .sourceBuild(let source, _),
-             .binaryDownload(let source, _):
-            return source.version
-        case .unknown:
-            return nil
-        }
-    }
-
-    var packageManagerType: PackageManagerType? {
-        switch self {
-        case .standardPackage(let source),
-             .sourceBuild(let source, _),
-             .binaryDownload(let source, _):
-            return source.type
-        case .unknown:
-            return nil
-        }
     }
 }
