+++ CodeEdit/Features/LSP/Registry/PackageManagers/GithubPackageManager.swift
@@ -10,7 +10,7 @@ import Foundation
 final class GithubPackageManager: PackageManagerProtocol {
     private let installationDirectory: URL
 
-    internal let shellClient: ShellClient
+    let shellClient: ShellClient
 
     init(installationDirectory: URL) {
         self.installationDirectory = installationDirectory
