+++ CodeEdit/Features/LSP/Registry/PackageManagers/GolangPackageManager.swift
@@ -9,7 +9,8 @@ import Foundation
 
 final class GolangPackageManager: PackageManagerProtocol {
     private let installationDirectory: URL
-    internal let shellClient: ShellClient
+
+    let shellClient: ShellClient
 
     init(installationDirectory: URL) {
         self.installationDirectory = installationDirectory
