+++ CodeEdit/Features/LSP/Registry/RegistryManagerError.swift
@@ -1,0 +1,15 @@
+//
+//  RegistryManagerError.swift
+//  CodeEdit
+//
+//  Created by Abe Malla on 5/12/25.
+//
+
+import Foundation
+
+enum RegistryManagerError: Error {
+    case invalidResponse(statusCode: Int)
+    case downloadFailed(url: URL, error: Error)
+    case maxRetriesExceeded(url: URL, lastError: Error)
+    case writeFailed(error: Error)
+}
